requirements.txt
streamlit
tensorflow
ultralytics
torch
pillow
numpy
